requirements.txt
mcp[cli]>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
lxml>=5.0.0
